created main functions

+++ src/thaw_slump_segmentation/scripts/process_02_inference.py
@@ -20,7 +20,6 @@ from ..postprocessing import (
     load_and_parse_vector,
     run_inference,
     update_DEM2,
-    print_processing_stats
 )
 
 
@@ -63,8 +62,17 @@ def process_02_inference(
     # get df for preprocessing
     df_final = pd.concat(df_processing_status_list).drop_duplicates()
 
-    # print basic information and return numbers
-    total_images, preprocessed_images, preprocessing_images, finished_images = print_processing_stats(df_final)
+    # TODO: move to function
+    # print basic information
+    total_images = int(len(df_final))
+    preprocessed_images = int(df_final.preprocessed.sum())
+    preprocessing_images = int(total_images - preprocessed_images)
+    finished_images = int(df_final.inference_finished.sum())
+    print(f'Number of images: {total_images}')
+    print(f'Number of preprocessed images: {preprocessed_images}')
+    print(f'Number of images for preprocessing: {preprocessing_images}')
+    print(f'Number of images for inference: {preprocessed_images - finished_images}')
+    print(f'Number of finished images: {finished_images}')
 
     # TODO: images with processing status True but Inference False are crappy
 
@@ -88,13 +96,15 @@ def process_02_inference(
     df_preprocess = df_final[df_final.preprocessed == False]
     print(f'Number of images to preprocess: {len(df_preprocess)}')
 
+    # TODO make better check
     # Cleanup processing directories to avoid incomplete processing
     input_dir_dslist = list((processing_dir / 'input').glob('*'))
     if len(input_dir_dslist) > 0:
         print(f"Cleaning up {(processing_dir / 'input')}")
         for d in input_dir_dslist:
-            print('Delete', d)
-            shutil.rmtree(d)
+            if len(list(d.glob('*'))) < 4:
+                print('Delete', d)
+                shutil.rmtree(d)
     else:
         print('Processing directory is ready, nothing to do!')
